adds multiple clients and executionStats
Adds multiple clients in a `MongoClients` struct that is used in the
data layer. The constituent clients are created with URIs derived from
environmental variables. The version information for which client to
used is passed in and checked, and the proper client is selected.

Additionally, `executionStats` are now returned as part of aggregation
execute. The driver doesn't expose a mechanism for explain other than
`run_command`, and according to https://github.com/mongodb/specifications/blob/master/source/crud/crud.rst#q-a this is the proper way.